# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.ErrorResponse:
    properties:
      data: {}
      message:
        example: 发生错误
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  controller.SuccessResponse:
    properties:
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
  controller.TrainingJobDTO:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      image:
        example: pytorch:latest
        type: string
      name:
        example: 训练任务名称
        type: string
      project_id:
        example: 1
        type: integer
      status:
        example: running
        type: string
      updated_at:
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  controller.TrainingJobResponse:
    properties:
      data:
        $ref: '#/definitions/controller.TrainingJobDTO'
      message:
        example: Training Job created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  controller.TrainingJobsListData:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      training_jobs:
        items:
          $ref: '#/definitions/controller.TrainingJobDTO'
        type: array
    type: object
  controller.TrainingJobsResponse:
    properties:
      data:
        $ref: '#/definitions/controller.TrainingJobsListData'
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.Project:
    type: object
  model.TrainingJob:
    properties:
      args:
        description: JSON-encoded array of arguments
        type: string
      command:
        description: JSON-encoded array of commands
        type: string
      cpu_limit:
        description: CPU limit per container
        type: string
      created_at:
        type: string
      env:
        description: JSON-encoded environment variables
        type: string
      gpus_per_node:
        description: Number of GPUs per node
        type: integer
      id:
        type: integer
      image:
        type: string
      image_pull_policy:
        type: string
      master_replicas:
        description: Number of master replicas
        type: integer
      memory_limit:
        description: Memory limit per container
        type: string
      name:
        type: string
      namespace:
        type: string
      node_selector:
        description: JSON-encoded node selector
        type: string
      parameters:
        description: JSON string for parameters
        type: string
      project:
        $ref: '#/definitions/model.Project'
      project_id:
        type: integer
      restart_policy:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
      worker_replicas:
        description: Number of worker replicas
        type: integer
    type: object
...
  model.User:
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for MLcore-Engine.
  title: MLcore-Engine API
  version: "1.0"
paths:
...
  /pytorchtrain:
    post:
      consumes:
      - application/json
      description: Create a new Training Job with the provided details
      parameters:
      - description: Training Job details
        in: body
        name: training_job
        required: true
        schema:
          $ref: '#/definitions/model.TrainingJob'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TrainingJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create a new Training Job
      tags:
      - training
  /pytorchtrain/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Training Job by its ID
      parameters:
      - description: Training Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Delete a Training Job
      tags:
      - training
  /training/{id}:
    get:
      consumes:
      - application/json
      description: Get a Training Job by its ID
      parameters:
      - description: Training Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TrainingJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get a Training Job
      tags:
      - training
  /training/get-all:
    get:
      consumes:
      - application/json
      description: Get a paginated list of Training Jobs
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TrainingJobsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List Training Jobs
      tags:
      - training
...
swagger: "2.0"